requirements.txt
numpy==1.18.5
streamlit==0.69.0
plotly==4.0.0
seaborn==0.10.1
pandas_profiling==2.3.0
pandas==1.0.5
matplotlib==3.3.0
scikit_learn==0.24.1
torch==1.2.0


